requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
annoy==1.17.3
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
dataclasses-json==0.6.7
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
environs==9.5.0
exceptiongroup==1.2.2
fastapi==0.115.3
fastembed==0.4.1
filelock==3.16.1
flatbuffers==24.3.25
frontend==0.0.3
frozenlist==1.4.1
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
grpcio==1.63.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.1
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.2.16
langchain-community==0.2.17
langchain-core==0.2.41
langchain-nvidia-ai-endpoints==0.2.2
langchain-text-splitters==0.2.4
langsmith==0.1.137
lark==1.1.9
llama-index-agent-openai==0.1.7
llama-index-core==0.10.58
llama-index-embeddings-nvidia==0.1.4
llama-index-embeddings-openai==0.1.11
llama-index-llms-nvidia==0.1.4
llama-index-llms-openai==0.1.31
llama-index-llms-openai-like==0.1.3
llama-index-readers-file==0.1.30
llama-index-vector-stores-milvus==0.1.20
loguru==0.7.2
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.0
mdurl==0.1.2
milvus-lite==2.4.10
mmh3==4.1.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.10.0
nemoguardrails==0.10.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.19.2
openai==1.52.0
orjson==3.10.10
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pip-autoremove==0.10.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.2
py_rust_stemmers==0.1.3
pyarrow==17.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pymilvus==2.4.4
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pypdf==4.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.3
rpds-py==0.20.0
safetensors==0.4.5
shellingham==1.5.4
simpleeval==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.41.0
streamlit==1.38.0
striprtf==0.0.26
sympy==1.13.3
tenacity==8.5.0
tiktoken==0.8.0
tokenizers==0.20.1
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
transformers==4.45.2
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.32.0
watchdog==4.0.2
wcwidth==0.2.13
wrapt==1.16.0
XlsxWriter==3.2.0
yarl==1.16.0
